+++ tools/check_site_structure.php
@@ -43,6 +43,7 @@ $requiredRootTools = [
     'check_menu_overlay.php',
     'check_confirm_modal.php',
     'check_js_syntax.php',
+    'check_password_toggle.php',
 ];
 
 $errors = [];